Created Details Class and cleaned up SingletonClass. Updated Details Fragment to reflect changes to Singleton
